autopilot: replace ChanGraph pointer with interface
