+++ graph/db/graph_test.go
@@ -1092,7 +1092,7 @@ func TestGraphTraversalCacheable(t *testing.T) {
 	// Create a map of all nodes with the iteration we know works (because
 	// it is tested in another test).
 	nodeMap := make(map[route.Vertex]struct{})
-	err = graph.ForEachNode(
+	err = graph.forEachNode(
 		func(tx kvdb.RTx, n *models.LightningNode) error {
 			nodeMap[n.PubKeyBytes] = struct{}{}
 
@@ -1217,7 +1217,7 @@ func fillTestGraph(t require.TestingT, graph *ChannelGraph, numNodes,
 
 	// Iterate over each node as returned by the graph, if all nodes are
 	// reached, then the map created above should be empty.
-	err := graph.ForEachNode(
+	err := graph.forEachNode(
 		func(_ kvdb.RTx, node *models.LightningNode) error {
 			delete(nodeIndex, node.Alias)
 			return nil
@@ -1329,7 +1329,7 @@ func assertNumChans(t *testing.T, graph *ChannelGraph, n int) {
 
 func assertNumNodes(t *testing.T, graph *ChannelGraph, n int) {
 	numNodes := 0
-	err := graph.ForEachNode(
+	err := graph.forEachNode(
 		func(_ kvdb.RTx, _ *models.LightningNode) error {
 			numNodes++
 			return nil
